refactor: Move "GameMetadata" types into "types" folder
